+++ lib/Elastica/Connection/Strategy/CallbackStrategy.php
@@ -11,19 +11,19 @@ use Elastica\Exception\InvalidException;
 class CallbackStrategy implements StrategyInterface
 {
     /**
-     * @var Closure
+     * @var callable
      */
     protected $_callback;
 
     /**
-     * @param Closure $callback
+     * @param callable $callback
      *
      * @throws \Elastica\Exception\InvalidException
      */
     public function __construct($callback)
     {
         if (!self::isValid($callback)) {
-            throw new InvalidException(sprintf('Callback should be a Closure, %s given!', gettype($callback)));
+            throw new InvalidException(sprintf('Callback should be a callable, %s given!', gettype($callback)));
         }
 
         $this->_callback = $callback;
@@ -36,14 +36,16 @@ class CallbackStrategy implements StrategyInterface
      */
     public function getConnection($connections)
     {
-        return $this->_callback->__invoke($connections);
+        return call_user_func_array($this->_callback, array($connections));
     }
 
     /**
+     * @param callable $callback
+     *
      * @return bool
      */
     public static function isValid($callback)
     {
-        return is_object($callback) && ($callback instanceof \Closure);
+        return is_callable($callback);
     }
 }