+++ test/lib/Elastica/Test/Connection/Strategy/CallbackStrategyTest.php
@@ -33,10 +33,28 @@ class CallbackStrategyTest extends Base
      */
     public function testIsValid()
     {
-        $callback = function () {};
+        // closure is valid
+        $isValid = CallbackStrategy::isValid(function () {});
+        $this->assertTrue($isValid);
 
-        $isValid = CallbackStrategy::isValid($callback);
+        // object implementing __invoke
+        $isValid = CallbackStrategy::isValid(new CallbackStrategyTestHelper());
+        $this->assertTrue($isValid);
 
+        // static method as string
+        $isValid = CallbackStrategy::isValid('Elastica\Test\Connection\Strategy\CallbackStrategyTestHelper::getFirstConnectionStatic');
+        $this->assertTrue($isValid);
+
+        // static method as array
+        $isValid = CallbackStrategy::isValid(array('Elastica\Test\Connection\Strategy\CallbackStrategyTestHelper', 'getFirstConnectionStatic'));
+        $this->assertTrue($isValid);
+
+        // object method
+        $isValid = CallbackStrategy::isValid(array(new CallbackStrategyTestHelper(), 'getFirstConnectionStatic'));
+        $this->assertTrue($isValid);
+
+        // function name
+        $isValid = CallbackStrategy::isValid('array_pop');
         $this->assertTrue($isValid);
     }
 
@@ -45,10 +63,10 @@ class CallbackStrategyTest extends Base
      */
     public function testFailIsValid()
     {
-        $callback = new \stdClass();
+        $isValid = CallbackStrategy::isValid(new \stdClass());
+        $this->assertFalse($isValid);
 
-        $isValid = CallbackStrategy::isValid($callback);
-
+        $isValid = CallbackStrategy::isValid('array_pop_pop_pop_pop_pop_pop');
         $this->assertFalse($isValid);
     }