+++ test/lib/Elastica/Test/Connection/Strategy/CallbackStrategyTestHelper.php
@@ -1,0 +1,20 @@
+<?php
+namespace Elastica\Test\Connection\Strategy;
+
+class CallbackStrategyTestHelper
+{
+    public function __invoke($connections)
+    {
+        return $connections[0];
+    }
+
+    public function getFirstConnection($connections)
+    {
+        return $connections[0];
+    }
+
+    public static function getFirstConnectionStatic($connections)
+    {
+        return $connections[0];
+    }
+}